Revert "Move esphome gatt services cache to be per device" #81265
This reverts the services cache to be per device and restores the
integration level cache. The reason it was previously unstable was
because of an issue with the esphome implementation that accidentally
mixed services between devices
